+++ ecmascript/transforms/compat/tests/es2015_function_name.rs
@@ -146,115 +146,6 @@ test!(
 //    return components_Link_extends.apply(this, arguments); }"
 //);
 
-// function_name_function_collision
-test!(
-    ignore,
-    syntax(),
-    |_| chain!(
-        resolver(),
-        decorators(decorators::Config {
-            legacy: true,
-            ..Default::default()
-        }),
-        classes(),
-        function_name(),
-    ),
-    function_name_function_collision,
-    r#"
-function f() {
-  f;
-}
-
-{
-  let obj = {
-    f: function () {
-      f;
-    }
-  };
-}
-
-(function b() {
-  var obj = {
-    b: function () {
-      b;
-    }
-  };
-
-  function commit(b) {
-    b();
-  }
-});
-
-"#,
-    r#"
-function _f() {
-  _f;
-}
-
-{
-  let obj = {
-    f: function f() {
-      _f;
-    }
-  };
-}
-
-(function _b() {
-  var obj = {
-    b: function b() {
-      _b;
-    }
-  };
-
-  function commit(b) {
-    b();
-  }
-});
-
-"#
-);
-
-// function_name_collisions
-test!(
-    syntax(),
-    |_| chain!(
-        resolver(),
-        decorators(decorators::Config {
-            legacy: true,
-            ..Default::default()
-        }),
-        classes(),
-        function_name(),
-    ),
-    function_name_collisions,
-    r#"
-var obj = {
-  search: function({search}) {
-    console.log(search);
-  }
-};
-
-function search({search}) {
-  console.log(search);
-}
-
-"#,
-    r#"
-var obj = {
-  search: function search({
-    search: search1
-  }) {
-    console.log(search1);
-  }
-};
-
-function search({ search: search1 }) {
-    console.log(search1);
-}
-
-"#
-);
-
 // issues_5004
 test!(
     syntax(),
